storage_compare/requirements.txt
annotated-types==0.7.0
certifi==2024.2.2
clickhouse-connect==0.7.11
Faker==25.2.0
lz4==4.3.3
mypy==1.10.0
mypy-extensions==1.0.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ruff==0.4.5
six==1.16.0
typing_extensions==4.12.0
urllib3==2.2.1
vertica-python==1.3.8
zstandard==0.22.0
